requirements.txt
accelerate==0.15.0
datasets==2.6.1
huggingface-hub==0.11.1
jiwer==2.5.1
librosa==0.9.2
espnet==202211
espnet_model_zoo==0.1.7
natsort==8.2.0
numpy==1.23.5
omegaconf==2.2.3
pandas==1.5.2
parallel-wavegan==0.5.5
pyctcdecode==0.4.0
soundfile==0.11.0
torch==1.13.0
torchaudio==0.13.0
tqdm==4.64.1
transformers==4.24.0
wandb==0.13.4